console.c/h: color setters for fg and bg now require inputs of type uint8_t

+++ src/console.c
@@ -104,7 +104,7 @@ void console_init()
 	console_putstring("\nconsole: initialized\n");
 }
 
-void console_set_fgcolor(int r, int g, int b)
+void console_set_fgcolor(uint8_t r, uint8_t g, uint8_t b)
 {
     r = console_verify_color_range(r);
     g = console_verify_color_range(g);
@@ -114,7 +114,7 @@ void console_set_fgcolor(int r, int g, int b)
     fgcolor.b = b;
 }
 
-void console_set_bgcolor(int r, int g, int b)
+void console_set_bgcolor(uint8_t r, uint8_t g, uint8_t b)
 {
     r = console_verify_color_range(r);
     g = console_verify_color_range(g);
@@ -124,7 +124,7 @@ void console_set_bgcolor(int r, int g, int b)
     bgcolor.b = b;
 }
 
-int console_verify_color_range(int x)
+uint8_t console_verify_color_range(uint8_t x)
 {
     if(x < 0) return 0;
     if(x > 255) return 255;
